Move the facility rt value pill into its own component
